+++ dbt/adapters/tidbcloudlake/impl.py
@@ -134,7 +134,16 @@ class TiDBCloudLakeAdapter(SQLAdapter):
         table = table_from_rows(
             table.rows,
             table.column_names,
-            text_only_columns=["table_schema", "table_name"],
+            text_only_columns=[
+                "table_schema",
+                "table_name",
+                "table_type",
+                "table_comment",
+                "table_owner",
+                "column_name",
+                "column_type",
+                "column_comment",
+            ],
         )
 
         schemas = frozenset(